ENG-2023 Move rtl util functions into the main testUtils module
