+++ test/test/testUtils.js
@@ -1,3 +1,17 @@
+import React from 'react';
+import { render, screen, within } from '@testing-library/react';
+import { combineReducers, createStore } from 'redux';
+import { Provider as StateProvider } from 'react-redux';
+import { reducer as formReducer } from 'redux-form';
+import { MemoryRouter, Route } from 'react-router-dom';
+import { IntlProvider } from 'react-intl';
+import configureMockStore from 'redux-mock-store';
+import thunk from 'redux-thunk';
+
+import enTranslations from 'locales/en';
+
+const { locale: enLocale, messages: enMessages } = enTranslations;
+
 export const errorResponse = (...messages) => ({
   errors: messages.map((message, code) => ({ code, message })),
 });
@@ -30,3 +44,94 @@ export const mockApi = ({
 };
 
 export const mockThunk = arg => () => () => new Promise(r => r(arg));
+
+const createMockStore = (state = {}) => {
+  const middlewares = [thunk];
+  const mockStore = configureMockStore(middlewares);
+  return mockStore(state);
+};
+
+export const renderWithWrapper = (ui, {
+  wrapperComp: WrapperComp, wrapperProps, ...renderOptions
+} = {}) => {
+  // eslint-disable-next-line react/prop-types
+  const Wrapper = ({ children }) => (
+    <WrapperComp {...wrapperProps}>
+      {children}
+    </WrapperComp>
+  );
+  return render(ui, { wrapper: Wrapper, ...renderOptions });
+};
+
+export const renderWithIntl = (ui, {
+  locale = enLocale, messages = enMessages, ...renderOptions
+} = {}) => renderWithWrapper(ui, {
+  wrapperComp: IntlProvider,
+  wrapperProps: { locale, messages },
+  ...renderOptions,
+});
+
+export const renderWithRouter = (ui, {
+  initialRoute = '/', path = '/', ...renderOptions
+} = {}) => renderWithWrapper(
+  (
+    <Route path={path}>{ui}</Route>
+  ), {
+    wrapperComp: MemoryRouter,
+    wrapperProps: { initialEntries: [initialRoute] },
+    ...renderOptions,
+  },
+);
+
+export const renderWithState = (ui, {
+  state, store = createMockStore(state), ...renderOptions
+} = {}) => renderWithWrapper(ui, {
+  wrapperComp: StateProvider,
+  wrapperProps: { store },
+  ...renderOptions,
+});
+
+export const renderWithIntlAndRouter = (ui, {
+  locale = enLocale, messages = enMessages, initialRoute = '/', path = '/', ...renderOptions
+} = {}) => renderWithRouter(
+  <IntlProvider locale={locale} messages={messages}>{ui}</IntlProvider>,
+  { initialRoute, path, ...renderOptions },
+);
+
+export const renderWithIntlAndState = (ui, {
+  locale = enLocale, messages = enMessages, state, store = createMockStore(state), ...renderOptions
+} = {}) => renderWithState(
+  <IntlProvider locale={locale} messages={messages}>{ui}</IntlProvider>,
+  { state, store, ...renderOptions },
+);
+
+export const renderWithIntlRouterState = (ui, {
+  locale = enLocale, messages = enMessages, initialRoute = '/', path = '/',
+  state, store = createMockStore(state), ...renderOptions
+} = {}) => renderWithState(
+  <MemoryRouter initialEntries={[initialRoute]}>
+    <IntlProvider locale={locale} messages={messages}>
+      <Route path={path}>{ui}</Route>
+    </IntlProvider>
+  </MemoryRouter>,
+  { state, store, ...renderOptions },
+);
+
+/**
+ * This requires the `ui` to have a form with an `aria-label` attribute defined.
+ */
+export const setupForm = (formName, formValues, renderWithStateFunc, ui) => {
+  const state = {
+    form: {
+      [formName]: {
+        values: formValues,
+      },
+    },
+  };
+  const store = createStore(combineReducers({
+    form: formReducer,
+  }), state);
+  const utils = renderWithStateFunc(ui, { store });
+  const formView = within(screen.getByRole('form'));
+  return { formView, formValues, ...utils };
+};
