+++ test/ui/email-config/EditEmailSenderFormContainer.test.js
@@ -3,7 +3,7 @@ import { fireEvent, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import * as reactRedux from 'react-redux';
 
-import { renderWithIntlRouterState, setupForm } from 'test/rtlTestUtils';
+import { renderWithIntlRouterState, setupForm } from 'test/testUtils';
 import userEvent from '@testing-library/user-event';
 import EditEmailSenderFormContainer from 'ui/email-config/EditEmailSenderFormContainer';
